requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
python-dotenv
pydantic
bcrypt
python-jose[cryptography]
httpx
tenacity
